Generate the enum OkModelingCmdResponse from the output structs
This previously used a declarative macro to automate some of the boilerplate. But it still required engineers to remember to update that enum, and to add a new line to its macro.

I've replaced that dec-macro with a proc-macro that wraps an entire module, where the output/response structs are defined. It generates an enum with one variant for each such struct.
