+++ modeling-cmds-macros/src/lib.rs
@@ -3,6 +3,7 @@
 mod modeling_cmd_enum;
 mod modeling_cmd_output;
 mod modeling_cmd_variant;
+mod ok_modeling_cmd_response_enum;
 
 use proc_macro::TokenStream;
 use syn::{DeriveInput, ItemMod};
@@ -42,4 +43,11 @@ pub fn derive_modeling_cmd_output(input: TokenStream) -> TokenStream {
     // For comments, see `derive_modeling_cmd_output_empty`.
     let input: DeriveInput = syn::parse2(input.into()).unwrap();
     TokenStream::from(modeling_cmd_output::derive(input))
+}
+
+/// Generates the OkModelingCmdResponse enum from all its variants.
+#[proc_macro]
+pub fn define_ok_modeling_cmd_response_enum(item: TokenStream) -> TokenStream {
+    let input: ItemMod = syn::parse2(item.into()).unwrap();
+    TokenStream::from(ok_modeling_cmd_response_enum::generate(input))
 }
